requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
streamlit>=1.28.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
python-multipart>=0.0.6
python-dateutil>=2.8.0
requests>=2.31.0

